added total hours calculator

+++ database/migrations/2016_04_09_211411_create_attendances_table.php
@@ -18,7 +18,7 @@ class CreateAttendancesTable extends Migration
             $table->date('date');
             $table->timestamp('timeIn');
             $table->timestamp('timeOut');
-            $table->decimal('actualHours', 2, 2);
+            $table->float('actualHours');
             $table->boolean('isHoliday')->default(false);
             $table->integer('leaveStatus');
             $table->timestamps();
